add smoke test for freqxlator and scanner.
