+++ .github/workflows/test.yml
@@ -7,7 +7,9 @@ jobs:
     - uses: actions/checkout@v2
     - name: static tests
       run: |
-        sudo apt-get update && sudo apt-get install -qy gcc git g++ libev-dev python3-dev python3-pip && \
+        sudo add-apt-repository ppa:gnuradio/gnuradio-releases-3.9 && \
+            sudo apt-get update && \
+            sudo apt-get install -qy gcc git g++ libev-dev python3-dev python3-pip gnuradio && \
         sudo pip3 install -r requirements.txt -r scan-requirements.txt -r tests/requirements.txt -r specgram-requirements.txt && \
         pip3 install codecov && \
         export PYTHONPATH=$PWD/gamutrf:$PYTHONPATH && \
